refactor: move 16 page components from components/ to pages/
Move page-level components that serve as route targets from components/
to pages/ directory for clearer project structure:

- LoginPage, VerifyEmailPage, ResetPasswordPage (auth)
- ProfilePage, BillingDashboard, TeamPage, HistoryPage (core)
- CaseAnalysisPage, DecisionsSearchPage, CourtPracticeAnalysisPage (legal)
- LegislationMonitoringPage, LegislationStatisticsPage (legislation)
- LegalInitiativesPage, VotingAnalysisPage (parliament)
- LegalCodesLibraryPage, HistoricalAnalysisPage (reference)

Updated all router imports and fixed relative paths in ProfilePage,
BillingDashboard, and TeamPage. Moved test file accordingly.
