+++ lexwebapp/src/pages/CaseAnalysisPage.tsx
