+++ lexwebapp/src/pages/CourtPracticeAnalysisPage.tsx
