+++ lexwebapp/src/pages/HistoricalAnalysisPage.tsx
