+++ lexwebapp/src/pages/HistoryPage.tsx
