+++ lexwebapp/src/pages/VotingAnalysisPage.tsx
